STYLE: Follow cmake naming conventions
The previous commit provided an maximally
backwards compatible naming for this
package, but it did not conform to the
general naming practice that the name
of the package should be used for
the find_package() calls.

It appears that the name of this package
consists of all capital letters ISMRMRD
and it would be most consistent for external
projects to look for build configuration
settings using -DISMRMRD_DIR:PATH=<path> from
the command line, and find_package(ISMRMRD REQUIRED)
from the CMakeLists.txt file.