COMP: Disable unnecessary enhanced version checking
Cmake provides some advanced version checking features
with automatic macros, but these appear to be causing
the travis build some issues.

This is not strictly needed for this patch set to
be considered complete, so removing these until the
travis issues can be identified more clearly and resolved.

NOTE: A mac and a unix build on local machines did not
have issues with this feature.