Fix erroneous use of float_t in recon_cartesian_2d (#164)
The call to memset uses the type float_t for calculating the area to
clear. However, the actual size of the underlying allocation is based on
the type float, as a result of how the Image is defined a few lines
above. In general, float_t and float are not necessarily the same. For
example, on x86 with -mfpmath=sse+387, float_t becomes long double; on
other architectures these types may differ generally. When that is the
case, the memset would perform an out-of-bounds access.

Fix to directly use sizeof(float).