Fix undefined behaviour (#166)
* Add DLL symbol export for function ismrmrd_sizeof_data_type.

Without this export, test_ismrmrd.exe isn't able to build.

* Fix undefined behaviour

The behaviour is undefined, if memcpy use an invalid or null pointer as source. Fix this by using memcpy only, if the data block of ndarray isn't NULL, otherwise set the destination data block to NULL.

* Fix undefined behaviour

The pointer carrp is not valid after calling ismrmrd_free_ndarray, therefore the result of the check if carrp->data is NULL is undefined and not needed.